Refactor imports in content_manager and reference files; add TableItem and VectorItem classes
