+++ src/kapipy/vector_item.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass, field
-from typing import List, Optional, Dict, Any, Union, Protocol, runtime_checkable
+from typing import List, Optional, Dict, Any, Union
 from datetime import datetime
 import logging
 
@@ -8,7 +8,6 @@ from .data_classes import BaseItem, WFS, VectorItemData
 from .conversion import (
     get_default_output_format,
     geojson_to_gdf,
-    json_to_df,
     geojson_to_sdf,
     bbox_sdf_into_cql_filter,
     geom_sdf_into_cql_filter,
@@ -416,157 +415,3 @@ class VectorItem(BaseItem, WFS):
 
         return f"Item id: {self.id}, type_: {self.type_}, title: {self.title}"
 
-
-@dataclass
-class TableItem(BaseItem, WFS):
-
-    def query_json(self, cql_filter: str = None, **kwargs: Any) -> dict:
-        """
-        Executes a WFS query on the item and returns the result as JSON.
-
-        Parameters:
-            cql_filter (str, optional): The CQL filter to apply to the query.
-            **kwargs: Additional parameters for the WFS query.
-
-        Returns:
-            dict: The result of the WFS query in JSON format.
-        """
-        logger.debug(f"Executing WFS query for item with id: {self.id}")
-
-        result = download_wfs_data(
-            url=self._wfs_url,
-            api_key=self._gis._api_key,
-            typeNames=f"{self.type_}-{self.id}",
-            cql_filter=cql_filter,
-            **kwargs,
-        )
-
-        return result
-
-    def query(self, cql_filter: str = None, **kwargs: Any) -> dict:
-        """
-        Executes a WFS query on the item and returns the result as a DataFrame.
-
-        Parameters:
-            cql_filter (str, optional): The CQL filter to apply to the query.
-            **kwargs: Additional parameters for the WFS query.
-
-        Returns:
-            pandas.DataFrame: The result of the WFS query as a DataFrame.
-        """
-        logger.debug(f"Executing WFS query for item with id: {self.id}")
-
-        result = self.query_json(cql_filter=cql_filter, **kwargs)
-
-        df = json_to_df(result, fields=self.data.fields)
-        return df
-
-    def get_changeset_json(
-        self, from_time: str, to_time: str = None, cql_filter: str = None, **kwargs: Any
-    ) -> dict:
-        """
-        Retrieves a changeset for the item in JSON format.
-
-        Parameters:
-            from_time (str): The start time for the changeset query, ISO format (e.g., "2015-05-15T04:25:25.334974").
-            to_time (str, optional): The end time for the changeset query, ISO format. If not provided, the current time is used.
-            cql_filter (str, optional): The CQL filter to apply to the changeset query.
-            **kwargs: Additional parameters for the WFS query.
-
-        Returns:
-            dict: The changeset data in JSON format.
-
-        Raises:
-            ValueError: If the item does not support changesets.
-        """
-
-        if not self.supports_changesets:
-            logger.error(f"Item with id: {self.id} does not support changesets.")
-            raise ValueError("This item does not support changesets.")
-
-        if to_time is None:
-            to_time = datetime.now().isoformat()
-        logger.debug(
-            f"Fetching changeset for item with id: {self.id} from {from_time} to {to_time}"
-        )
-
-        viewparams = f"from:{from_time};to:{to_time}"
-
-        result = download_wfs_data(
-            url=self._wfs_url,
-            api_key=self._gis._api_key,
-            typeNames=f"{self.type_}-{self.id}-changeset",
-            viewparams=viewparams,
-            cql_filter=cql_filter,
-            **kwargs,
-        )
-
-        return result
-
-    def get_changeset(
-        self, from_time: str, to_time: str = None, cql_filter: str = None, **kwargs: Any
-    ) -> dict:
-        """
-        Retrieves a changeset for the item and returns it as a DataFrame.
-
-        Parameters:
-            from_time (str): The start time for the changeset query, ISO format (e.g., "2015-05-15T04:25:25.334974").
-            to_time (str, optional): The end time for the changeset query, ISO format. If not provided, the current time is used.
-            cql_filter (str, optional): The CQL filter to apply to the changeset query.
-            **kwargs: Additional parameters for the WFS query.
-
-        Returns:
-            pandas.DataFrame: The changeset data as a DataFrame.
-        """
-
-        result = self.get_changeset_json(
-            from_time=from_time, to_time=to_time, cql_filter=cql_filter, **kwargs
-        )
-
-        df = json_to_df(result, fields=self.data.fields)
-        return df
-
-    def export(
-        self,
-        export_format: str,
-        poll_interval: int = None,
-        timeout: int = None,
-        **kwargs: Any,
-    ) -> JobResult:
-        """
-        Exports the item in the specified format.
-
-        Parameters:
-            export_format (str): The format to export the item in.
-            poll_interval (int, optional): The interval in seconds to poll the export job status. Default is 10 seconds.
-            timeout (int, optional): The maximum time in seconds to wait for the export job to complete. Default is 600 seconds (10 minutes).
-            **kwargs: Additional parameters for the export request.
-
-        Returns:
-            JobResult: A JobResult instance containing the export job details.
-
-        Raises:
-            ValueError: If export validation fails.
-        """
-
-        logger.debug(f"!! Exporting item with id: {self.id} in format: {export_format}")
-
-        job_result = self._gis.content.export(
-            itm=self,
-            export_format=export_format,
-            poll_interval=poll_interval,
-            timeout=timeout,
-            **kwargs,
-        )
-
-        logger.debug(
-            f"Export job created for item with id: {self.id}, job id: {job_result.id}"
-        )
-        return job_result
-
-    def __str__(self) -> None:
-        """
-        User friendly string of a base item.
-        """
-
-        return f"Item id: {self.id}, type_: {self.type_}, title: {self.title}"
